Refactor: 편지 목록 조회를 LetterList에서 LettersItem으로 이름 변경, 컴포넌트로 쪼개기기
